+++ src/js/renderers/hls.js
@@ -210,10 +210,14 @@ const HlsNativeRenderer = {
 								break;
 							case 'networkError':
 								if (data.details === 'manifestLoadError') {
-									if (index < total) {
+									if (index < total && mediaFiles[(index + 1)] !== undefined) {
 										node.setSrc(mediaFiles[index++].src);
 										node.load();
 										node.play();
+									} else {
+										const message = 'Network error';
+										mediaElement.generateError(message, mediaFiles);
+										console.error(message);
 									}
 								} else {
 									const message = 'Network error';